gps abstract class extending iplookup, location and search classes  implemented
