+++ src/Objects/GPS/Search.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace GrigoryGerasimov\Weather\Objects\GPS;
+
+final readonly class Search extends Gps
+{
+    public function __construct(
+        private \stdClass $search
+    ) {
+        parent::__construct($search);
+    }
+
+    public function getId(): ?int
+    {
+        return $this->search->id;
+    }
+
+    public function getUrl(): ?string
+    {
+        return $this->search->url;
+    }
+}
